use range_utils.py instead of range_lib.py
