+++ tb_plugin/torch_tb_profiler/profiler/gpu_metrics_parser.py
@@ -1,4 +1,4 @@
-from .range_lib import *
+from .range_utils import *
 from .trace import EventTypes
 from .. import utils
