--- tb_plugin/torch_tb_profiler/profiler/range_lib.py
@@ -1,144 +1,0 @@
-from enum import IntEnum
-
-
-EndpointTypes = IntEnum('EndpointTypes', ['START', 'END'], start=0)
-
-
-class EndPoint(object):
-    def __init__(self, ep_time, ep_pt_type, ep_value):
-        self.time = ep_time
-        self.pt_type = ep_pt_type
-        self.value = ep_value
-
-
-# src_ranges: item of ((start_time, end_time), value)
-def merge_ranges_with_value(src_ranges):
-    merged_ranges = []
-    if len(src_ranges) > 0:
-        # Build tuple of (time, type, value)
-        endpoints = []
-        for r in src_ranges:
-            endpoints.append(EndPoint(r[0][0], EndpointTypes.START, r[1]))
-            endpoints.append(EndPoint(r[0][1], EndpointTypes.END, r[1]))
-        endpoints.sort(key=lambda x: [x.time, int(x.pt_type)])  # Make START in front of END if equal on time.
-
-        last_endpoint_time = endpoints[0].time
-        last_value = endpoints[0].value
-        for i in range(1, len(endpoints)):
-            ep = endpoints[i]
-            if ep.time > last_endpoint_time and last_value > 0.0:
-                approximated_sm_efficiency = min(last_value, 1.0)
-                merged_ranges.append(((last_endpoint_time, ep.time), approximated_sm_efficiency))
-            last_endpoint_time = ep.time
-            if ep.pt_type == EndpointTypes.START:
-                last_value += ep.value
-            else:
-                last_value -= ep.value
-
-    return merged_ranges
-
-
-def merge_ranges(src_ranges, is_sorted=False):
-    merged_ranges = []
-    if len(src_ranges) > 0:
-        if not is_sorted:
-            src_ranges.sort(key=lambda x: x[0])
-        src_id = 0
-        merged_ranges.append(
-            (src_ranges[src_id][0], src_ranges[src_id][1]))
-        for src_id in range(1, len(src_ranges)):
-            dst_id = len(merged_ranges) - 1
-            if src_ranges[src_id][1] > merged_ranges[dst_id][1]:
-                if src_ranges[src_id][0] <= merged_ranges[dst_id][1]:
-                    merged_ranges[dst_id] = (merged_ranges[dst_id][0], src_ranges[src_id][1])
-                else:
-                    merged_ranges.append(
-                        (src_ranges[src_id][0], src_ranges[src_id][1]))
-    return merged_ranges
-
-
-def pop_list(range_list, index):
-    next_index = index + 1
-    if next_index >= len(range_list):
-        return None, len(range_list)
-    next_item = range_list[next_index]
-    return next_item, next_index
-
-
-def subtract_ranges_lists(range_list1, range_list2):
-    range_list_dst = []
-    if len(range_list1) == 0:
-        return range_list_dst
-    if len(range_list2) == 0:
-        range_list_dst = list(range_list1)
-        return range_list_dst
-    r1 = range_list1[0]
-    r2 = range_list2[0]
-    i1 = i2 = 0
-    while i1 < len(range_list1):
-        if i2 == len(range_list2):
-            range_list_dst.append(r1)
-            r1, i1 = pop_list(range_list1, i1)
-        elif r2[1] <= r1[0]:
-            r2, i2 = pop_list(range_list2, i2)
-        elif r2[0] <= r1[0] and r2[1] < r1[1]:
-            r1 = (r2[1], r1[1])
-            r2, i2 = pop_list(range_list2, i2)
-        elif r2[0] <= r1[0]:
-            assert (r2[1] >= r1[1])
-            r2 = (r1[1], r2[1])
-            r1, i1 = pop_list(range_list1, i1)
-        elif r2[0] < r1[1]:
-            assert (r2[0] > r1[0])
-            range_list_dst.append((r1[0], r2[0]))
-            r1 = (r2[0], r1[1])
-        else:
-            assert (r2[0] >= r1[1])
-            range_list_dst.append(r1)
-            r1, i1 = pop_list(range_list1, i1)
-    return range_list_dst
-
-
-def intersection_ranges_lists(range_list1, range_list2):
-    range_list_dst = []
-    if len(range_list1) == 0 or len(range_list2) == 0:
-        return range_list_dst
-    r1 = range_list1[0]
-    r2 = range_list2[0]
-    i1 = i2 = 0
-    while i1 < len(range_list1):
-        if i2 == len(range_list2):
-            break
-        elif r2[1] <= r1[0]:
-            r2, i2 = pop_list(range_list2, i2)
-        elif r2[0] <= r1[0] and r2[1] < r1[1]:
-            assert (r2[1] > r1[0])
-            range_list_dst.append((r1[0], r2[1]))
-            r1 = (r2[1], r1[1])
-            r2, i2 = pop_list(range_list2, i2)
-        elif r2[0] <= r1[0]:
-            assert (r2[1] >= r1[1])
-            range_list_dst.append(r1)
-            r2 = (r1[1], r2[1])
-            r1, i1 = pop_list(range_list1, i1)
-        elif r2[1] < r1[1]:
-            assert (r2[0] > r1[0])
-            range_list_dst.append(r2)
-            r1 = (r2[1], r1[1])
-            r2, i2 = pop_list(range_list2, i2)
-        elif r2[0] < r1[1]:
-            assert (r2[1] >= r1[1])
-            range_list_dst.append((r2[0], r1[1]))
-            r2 = (r1[1], r2[1])
-            r1, i1 = pop_list(range_list1, i1)
-        else:
-            assert (r2[0] >= r1[1])
-            r1, i1 = pop_list(range_list1, i1)
-    return range_list_dst
-
-
-def get_ranges_sum(ranges):
-    sum = 0
-    for range in ranges:
-        sum += (range[1] - range[0])
-    return sum
